Always pass absolute cert path to Pip.
Previously, if the `--cert` path supplied was relative, `--isolated`
builds would fail when the --cert was passed to Pip via env var and the
isolated build needed to resolve `setup_requires` using the cert since
the isolated build would have a fresh new CWD. Add a failing test for
this and fix.

Fixes #1537
