+++ tests/integration/conftest.py
@@ -4,17 +4,18 @@
 from __future__ import absolute_import
 
 import os
-import shutil
 import subprocess
+from contextlib import contextmanager
+from textwrap import dedent
 
 import pytest
 
-from pex.common import atomic_directory, safe_mkdir
-from pex.testing import make_env, run_pex_command
+from pex.common import atomic_directory, temporary_dir
+from pex.testing import PY38, ensure_python_venv, make_env, run_pex_command
 from pex.typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from typing import Any
+    from typing import Any, Callable, ContextManager, Iterator, Optional, Tuple
 
 
 @pytest.fixture(scope="session")
@@ -60,3 +61,70 @@ def pex_bdist(
     wheels = os.listdir(wheels_dir)
     assert 1 == len(wheels)
     return os.path.join(wheels_dir, wheels[0])
+
+
+@pytest.fixture
+def tmp_workdir():
+    # type: () -> Iterator[str]
+    cwd = os.getcwd()
+    with temporary_dir() as tmpdir:
+        os.chdir(tmpdir)
+        try:
+            yield os.path.realpath(tmpdir)
+        finally:
+            os.chdir(cwd)
+
+
+@pytest.fixture(scope="module")
+def mitmdump():
+    # type: () -> Tuple[str, str]
+    python, pip = ensure_python_venv(PY38)
+    subprocess.check_call([pip, "install", "mitmproxy==5.3.0"])
+    mitmdump = os.path.join(os.path.dirname(python), "mitmdump")
+    return mitmdump, os.path.expanduser("~/.mitmproxy/mitmproxy-ca-cert.pem")
+
+
+@pytest.fixture
+def run_proxy(mitmdump, tmp_workdir):
+    # type: (Tuple[str, str], str) -> Callable[[Optional[str]], ContextManager[Tuple[int, str]]]
+    messages = os.path.join(tmp_workdir, "messages")
+    addon = os.path.join(tmp_workdir, "addon.py")
+    with open(addon, "w") as fp:
+        fp.write(
+            dedent(
+                """\
+                from mitmproxy import ctx
+        
+                class NotifyUp:
+                    def running(self) -> None:
+                        port = ctx.master.server.address[1]
+                        with open({msg_channel!r}, "w") as fp:
+                            print(str(port), file=fp)
+        
+                addons = [NotifyUp()]
+                """.format(
+                    msg_channel=messages
+                )
+            )
+        )
+
+    @contextmanager
+    def _run_proxy(
+        proxy_auth=None,  # type: Optional[str]
+    ):
+        # type: (...) -> Iterator[Tuple[int, str]]
+        os.mkfifo(messages)
+        proxy, ca_cert = mitmdump
+        args = [proxy, "-p", "0", "-s", addon]
+        if proxy_auth:
+            args.extend(["--proxyauth", proxy_auth])
+        proxy_process = subprocess.Popen(args)
+        try:
+            with open(messages, "r") as fp:
+                port = int(fp.readline().strip())
+                yield port, ca_cert
+        finally:
+            proxy_process.kill()
+            os.unlink(messages)
+
+    return _run_proxy
